Adding spotbugs to the build.
Spotbugs catches at least one real bug (this patch contains the correction).
It looks like a good idea to enable it.

Note that it probably increases the build time slightly (no metric at this
point).

Another patch will add some more configuration to check non-nullity.

+++ src/main/java/org/wikimedia/search/extra/latency/SearchLatencyListener.java
@@ -140,7 +140,7 @@ public class SearchLatencyListener extends AbstractLifecycleComponent implements
         }
 
         synchronized double getMillisAtPercentile(double percentile) {
-            return current.getValueAtPercentile(percentile) / TimeValue.NSEC_PER_MSEC;
+            return current.getValueAtPercentile(percentile) / ((double)TimeValue.NSEC_PER_MSEC);
         }
 
         synchronized TimeValue getTimeValueAtPercentile(double percentile) {
